ai-browser-agent/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
pyyaml==6.0.1
playwright==1.41.0
anthropic==0.18.1
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
